Update checker tool
Results of MACS2 peak calling may differ
even with the same workflow inputs files.
It depends on the order of reads in BAM
file, which differ from run to run even
if it's sorted, especially by coordinates.
Assuming that the BAM file, which is used
as input for MACS2, is correct - the results
of MACS2 will be correct too, if it was
called with the same arguments. So to
check if MACS2 gives correct results, we
need to check only arguments, that were
used to run program